stock-research-system/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pymongo==4.10.1
motor==3.3.2
pydantic==2.5.0
httpx==0.25.1
python-multipart==0.0.6
websockets==12.0
langchain==0.0.350
langchain-openai==0.0.5
langgraph==0.0.30
tavily-python==0.3.0
redis==5.0.1
aioredis==2.0.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
yfinance==0.2.33
aiohttp==3.9.1
reportlab==4.0.7

# AI Enhancement Dependencies
langchain-anthropic==0.1.1
tiktoken>=0.5.1,<0.6.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
langchain-community==0.0.10

# Server-Sent Events (Phase 2)
sse-starlette==1.8.2

# Production Server (Phase 3)
gunicorn==21.2.0
# LLM Observability (Phase 3)
langfuse==2.6.3

# BigQuery Data Lake (Phase 4)
google-cloud-bigquery==3.14.0

# Advanced Analytics (Phase 4)
pandas==2.1.3
numpy==1.24.3
scipy==1.11.4
